Remove dependency on big
