+++ docker-compose-linux.yml
@@ -1,0 +1,75 @@
+version: '2'
+services:
+  zookeeper:
+    image: registry.banno-internal.com/zookeeper:3.4.8
+    ports:
+      - 2181:2181
+    network_mode: host
+
+  marathon:
+    image: mesosphere/marathon:v1.3.3
+    ports:
+      - 8080:8080
+    command: "--master zk://localhost:2181/mesos --zk zk://localhost:2181/marathon"
+    environment:
+      MARATHON_MESOS_ROLE: marathon
+    depends_on:
+      - zookeeper
+      - mesos-master
+      - mesos-agent1
+      - mesos-agent2
+    network_mode: host
+
+  mesos-master:
+    image: registry.banno-internal.com/mesos-master:1.0
+    ports:
+      - 5050:5050
+    environment:
+      MESOS_ZK: zk://localhost:2181/mesos
+      MESOS_LOG_DIR: "/var/log"
+      MESOS_WORK_DIR: "/mesos"
+      MESOS_QUORUM: 1
+      MESOS_CLUSTER: localhost
+    depends_on:
+      - zookeeper
+    network_mode: host
+
+  mesos-agent1:
+    image: registry.banno-internal.com/mesos-agent:1.0
+    command: mesos-agent --resources=cpus:4;mem:24576;disk:409600
+    ports:
+      - 5151:5151
+    environment:
+      MESOS_CONTAINERIZERS: docker,mesos
+      MESOS_DOCKER_STOP_TIMEOUT: 30secs
+      MESOS_LOG_DIR: "/var/log"
+      MESOS_MASTER: zk://localhost:2181/mesos
+      MESOS_WORK_DIR: "/mesos"
+      MESOS_PORT: '5151'
+      MESOS_HOSTNAME: localhost
+    volumes:
+      - "/sys:/sys"
+      - "/var/run/docker.sock:/var/run/docker.sock"
+    depends_on:
+      - zookeeper
+    network_mode: host
+
+  mesos-agent2:
+    image: registry.banno-internal.com/mesos-agent:1.0
+    command: mesos-agent --resources=ports:[41000-42000];cpus:4;mem:24576;disk:409600
+    ports:
+      - 5152:5152
+    environment:
+      MESOS_CONTAINERIZERS: docker,mesos
+      MESOS_DOCKER_STOP_TIMEOUT: 30secs
+      MESOS_LOG_DIR: "/var/log"
+      MESOS_MASTER: zk://localhost:2181/mesos
+      MESOS_WORK_DIR: "/mesos"
+      MESOS_PORT: '5152'
+      MESOS_HOSTNAME: localhost
+    volumes:
+      - "/sys:/sys"
+      - "/var/run/docker.sock:/var/run/docker.sock"
+    depends_on:
+      - zookeeper
+    network_mode: host
