+++ makefile
@@ -9,14 +9,22 @@ linux:
 osx:
 	GOOS=darwin GOARCH=386 go build -o bin/terraform-provider-marathon-osx .
 
-build: vet osx linux
+install:
 	go install .
 
-test: build
-	big inventory add marathon
-	big up -d marathon
-	sleep 5
-	TF_LOG=TRACE TF_LOG_PATH=./test-sh-tf.log TF_ACC=yes MARATHON_URL='http://dev.banno.com:8080' go test ./marathon -v
+os=$(shell uname)
+ifeq ($(os),Darwin)
+  docker_compose_file=docker-compose.yml
+else
+  docker_compose_file=docker-compose-linux.yml
+endif
 
-release:
+test: install
+	docker-compose -f $(docker_compose_file) up -d
+	sleep 5
+	TF_LOG=TRACE TF_LOG_PATH=./test-sh-tf.log TF_ACC=yes MARATHON_URL=http://dev.banno.com:8080 go test ./marathon -v
+	docker-compose -f $(docker_compose_file) kill
+	docker-compose -f $(docker_compose_file) rm -f
+
+release: vet linux osx
 	./bin/release.sh
